+++ src/components/pages/apps/matcalc/data/data-props-panel.tsx
@@ -1,44 +1,57 @@
-import {
-  Tabs,
-  TabsContent,
-  TabsList,
-  TabsTrigger,
-} from '@/components/shadcn/ui/themed/v2/tabs'
-import { useState } from 'react'
+import { BaseCol } from '@/components/layout/base-col'
+import { Tabs, TabsContent } from '@/components/shadcn/ui/themed/v2/tabs'
+import { useTabs } from '@/components/tabs/tab-provider'
+import { useEffect } from 'react'
 import { FilterPropsPanel } from './filter-props-panel'
 import { GroupingPropsPanel } from './grouping-props-panel'
+import { SidebarTabs } from './sidebar-tabs'
 
 // const TAB_CLS =
 //   'w-4.5 stroke-foreground/75 group-hover:stroke-foreground group-data-[selected=true]:stroke-foreground'
 
 export function DataPropsPanel() {
-  const [value, setValue] = useState('groups')
+  const { selectedTab, setTabs } = useTabs('matcalc-data-props-panel')
+  //const [value, setValue] = useState('groups')
+
+  useEffect(() => {
+    setTabs([
+      {
+        id: 'groups',
+        name: 'Groups',
+      },
+      {
+        id: 'filter',
+        name: 'Filter',
+      },
+    ])
+  }, [setTabs])
 
   return (
-    <Tabs
-      value={value}
-      onValueChange={setValue}
-      orientation="vertical"
-      className="flex flex-col grow text-xs pr-1"
-    >
-      <TabsContent value="groups" className="grow">
-        <GroupingPropsPanel />
-      </TabsContent>
+    <BaseCol className="grow">
+      <Tabs
+        value={selectedTab?.id ?? ''}
+        //onValueChange={setTab}
+        orientation="vertical"
+        className="grow"
+      >
+        <TabsContent value="groups" className="grow">
+          <GroupingPropsPanel />
+        </TabsContent>
 
-      <TabsContent value="filter" className="grow">
-        <FilterPropsPanel />
-      </TabsContent>
-      <TabsList className="gap-y-px">
-        <TabsTrigger value="groups" className="grow" variant="sidebar">
-          Groups
-        </TabsTrigger>
-        {/* <TabsTrigger value="genesets" className="grow" variant="sidebar">
-          Gene Sets
-        </TabsTrigger> */}
-        <TabsTrigger value="filter" className="grow" variant="sidebar">
-          Filter
-        </TabsTrigger>
-      </TabsList>
-    </Tabs>
+        <TabsContent value="filter" className="grow">
+          <FilterPropsPanel />
+        </TabsContent>
+        {/* <TabsList className="gap-y-px">
+          <TabsTrigger value="groups" className="grow" variant="sidebar">
+            Groups
+          </TabsTrigger>
+
+          <TabsTrigger value="filter" className="grow" variant="sidebar">
+            Filter
+          </TabsTrigger>
+        </TabsList> */}
+      </Tabs>
+      <SidebarTabs id="matcalc-data-props-panel" />
+    </BaseCol>
   )
 }